Updated TCLogging to remove the hard log4j dependencies
 - Supports Log4j  by default if it is in the classpath, otherwise uses SLF4J
 - API break: had to remove TCLogger.setLevel() method and some other public methods on TCLogger
 - Provided implementations of TCLogger and TCLogging for both Log4j and SLF4J
 - Removed the JMX Appender (confirmed with @lorban)
 - Client Runtime only contains SLF4J api and ships with externalized log4j and slf4j-log4j

+++ common/src/main/java/com/tc/logging/Log4JAppenderToTCAppender.java
@@ -34,7 +34,7 @@ public class Log4JAppenderToTCAppender extends AppenderSkeleton {
   protected void append(LoggingEvent event) {
     ThrowableInformation throwableInformation = event.getThrowableInformation();
     Throwable t = (throwableInformation == null) ? null : throwableInformation.getThrowable();
-    appender.append(LogLevelImpl.fromLog4JLevel(event.getLevel()), event.getMessage(), t);
+    appender.append(LogLevelImpl.fromLog4JLevel(event.getLevel().toInt()), event.getMessage(), t);
   }
 
   @Override
